Change some callbacks to be called by setTimeout..
instead of direct call.
This is maybe more appropriate because setTimeout(callback, 0)
puts the callback to the message queue of Javascript runtime event loop,
so the callback does not stack upon the stack of the caller function and
also does not block if the callbacks are called a bunch in a row.